backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
mysql-connector-python==8.2.0
PyJWT==2.8.0
bcrypt==4.1.2
python-multipart==0.0.6
langchain==0.1.0
langchain-community==0.0.10
langchain-experimental==0.0.49
SQLAlchemy==2.0.23
PyMySQL==1.1.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.2
